Issue #249 - extracted class InmateDetails from test case
